# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API specification
  description: |
    This OpenAPI document describes the API for a messaging web application.
  version: "0.0.1"
...
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /conversations:
    get:
      tags: ["conversations"]
      summary: Retrive conversations
      description: List all the conversations of the users, sorted reverse chronologically.
      operationId: getMyConversations
      responses:
        "200":
          description: Successfully retrieved conversation list
          content:
            application/json:
              schema:
                type: array
                description: Array of conversation list items
                minItems: 0
                maxItems: 20
                items:
                  $ref: "#/components/schemas/conversation-info"
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
    post:
      tags: ["conversations"]
      summary: Create conversation (DM)
      description: Create a new one-to-one conversation with another user
      operationId: createConversation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                recipientName:
                  type: string
              required: [recipientName]
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/conversation-info"
        "404":
          description: Recipient not found
...
  /groups:
    post:
      tags: ["group"]
      summary: Create group
      description: Create a new group and specify the initial members
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: The name and members of the new group
              properties:
                name:
                  type: string
                  description: The name of the new group
                  minLength: 3
                  maxLength: 20
                  pattern: '^.*?$'
                initialMembers:
                  type: array
                  description: List of user IDs include in the group (excluding the creator)
                  minItems: 2 
                  # The group must has at least three person, otherwise its a conversation
                  maxItems: 100
                  items:
                    type: integer
              required:
                - name
                - initialMembers
        required: true
      responses:
        "201":
          description: Create a new group successfully, returns the new group ID
          content:
            application/json:
              schema:
                type: object
                description: The new group ID
                properties:
                  groupId:
                    type: integer
                    description: The ID of newly created group
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /groups/{groupId}/members:
    parameters:
      - name: groupId
        in: path
        required: true
        description: this is the group id
        schema:
          type: integer
          description: an incremental number 
    get:
      tags: ["group"]
      summary: Get group members
      description: Returns a list of users in the group
      operationId: getGroupMembers
      responses:
        "200":
          description: Successfully retrieved group members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user-info"
        "401":
          description: Unauthorized
        "404":
          description: Group not found
          
    post:
      tags: ["group"]
      summary: Add to group
      description: Group members can add other users to the group 
      operationId: addToGroup
      requestBody:
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/userIdsRequest'
        required: true
      responses:
        "200":
          description: Add user successfully
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "403":
          description: No permission to add members to this group
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "404":
          description: Group not found or some user IDs are invalid
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /groups/{groupId}/me:
    parameters:
      - name: groupId
        in: path
        required: true
        description: this is the group id
        schema:
          type: integer
          description: an incremental number 
          
    delete:
      tags: ["group"]
      summary: Leave group
      description: The current user exits the group
      operationId: leaveGroup
      responses:
        "200":
          description: Leave the group successfully
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "404":
          description: Group not found
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /groups/{groupId}/name:
    parameters:
      - name: groupId
        in: path
        required: true
        description: this is the group id
        schema:
          type: integer
          description: an incremental number 
          
    put:
      tags: ["group"]
      summary: Set group name
      operationId: setGroupName
      description: Updates the display name for an existing group
      requestBody:
        description: Group name
        content:
          application/json:
            schema:
              type: object
              description: The new group name info
              properties:
                newName:
                  type: string
                  description: The new name for the group
                  minLength: 3
                  maxLength: 20
                  pattern: '^.*?$'
              required:
                - newName
        required: true
      responses:
        "200":
          description: Update group name successfully
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "403":
          description: No permission to access
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /groups/{groupId}/photo:
    parameters:
      - name: groupId
        in: path
        required: true
        description: this is the group id
        schema:
          type: integer
          description: an incremental number 
          
    put:
      tags: ["group"]
      summary: Set group photo
      operationId: setGroupPhoto
      description: Uploads and sets a new profile pricture for the group
      requestBody:
        description: Group photo
        content:
          multipart/form-data:
            schema:
              type: object
              description: The group photo file
              properties:
                photo:
                  type: string
                  format: binary
                  description: The group photo file to upload
              required:
                - photo
        required: true
      responses:
        "200":
          description: Update group name successfully
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "403":
          description: No permission to access
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: WASAText-ID
      description: |
        Authentication is performed by passing the user's ID 
        in the Authorization header as a Bearer token
  schemas:
    ErrorResponse:
      type: object
      description: Standard error response format
      properties:
        error:
          type: string
          description: A brief description of error
          minLength: 0
          maxLength: 200
        code:
          type: integer
          description: Optional specific error code
      required: 
        - error
    user-info:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: string
      required: [id, name]
    conversation-info:
      type: object
      description: Information about a single conversation shown in the list
      properties:
        conversationId:
          type: integer
          description: The unique ID of the conversation
        name:
          type: string
          description: The display name of the conversation
          minLength: 1
          maxLength: 20
          pattern: '^.*?$'
        photoUrl:
          type: string
          format: url
          description: URL to the user/group photo
          pattern: '^https?://.+$'
        isGroup: 
          type: boolean
          description: True if the conversation is a group chat
        latestMessageTime:
          type: string
          format: date-time
          description: Timestamp of the latest message
        latestMessagePreview:
          type: string
          description: A short snippet of the latest message content
          minLength: 0
          maxLength: 99
        unreadCount:
          type: integer
          description: Number of unread messages in this conversation
          minimum: 0
      required:
        - conversationId
        - name
        - isGroup
        - latestMessageTime
        - latestMessagePreview
        - unreadCount
...
    userIdsRequest:
      type: object
      description: Request body containing a list of user IDs
      properties:
        userIds:
          type: array
          description: A list of user IDs
          minItems: 1
          maxItems: 50
          items:
            type: integer
      required: 
        - userIds
security:
  - bearerAuth: []